Add a way to enable/disable loading of included standard fonts.
Even when useSystemFonts is true, we will still default to
using the included fonts for symbol fonts since they are so
inconsistent across platforms.

I also added another flag `isInternalFont` to signify the font
was loaded from data. Whereas, `isStandardFont` is used
to signal if the font is a standard font, regardless of what's
used to show it.